requirements.txt
flask
openai==0.28
python-dotenv
flask-cors

